Oops

+++ src/main/java/cn/nukkit/network/protocol/types/ScorePacketEntry.java
@@ -7,7 +7,7 @@ public class ScorePacketEntry {
     public static final int TYPE_ENTITY = 2;
     public static final int TYPE_FAKE_PLAYER = 3;
 
-    public int scoreboardId;
+    public long scoreboardId;
     public String objectiveName;
     public int score;
